polymarket_bot/requirements.txt
py-clob-client==0.18.0
requests==2.31.0
python-dotenv==1.0.1
# web3==6.20.0   # uncomment if using on-chain features
